daemon: Check NodePort --device names and ifindices
Check that ifindex of a given device does not overflow uin16, and that
the device name does not contain special symbols used to parse in
bpf/init.sh.

The first restriction will go away once we introduce the surrogate id =>
ifindex BPF map. While the second - when we migrate bpf/init.sh to Go.

Signed-off-by: Martynas Pumputis <m@lambda.lt>