Endpoint for host
This commit adds a special Cilium endpoint on each node to represent the
node itself. It takes a reserved endpoint ID 4096 and the reserved
security ID 1. The special endpoint is created by
endpoint.CreateLocalNodeEndpoint on daemon startup if that endpoint wasn't
already restored.

When the datapath for the special endpoint is regenerated, it loads
bpf_host.o instead of bpf_lxc.o. bpf_host enforces policies for both
traffic to/from pods and traffic to/from the outside world. It follows
the same template generation path as traditional bpf_lxc files.

ENABLE_EXTRA_HOST_DEV was only defined in init.sh for the from_host path.
It is now defined from the Go side for all paths, and we check whether
from_host is true in the C code instead.

Signed-off-by: Paul Chaignon <paul@cilium.io>